pyproject.toml
[project]
name = "document-formatter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.2",
    "docx2txt>=0.9",
    "fastapi>=0.129.2",
    "gunicorn>=25.1.0",
    "jinja2>=3.1.6",
    "llama-index>=0.14.15",
    "mistralai>=1.12.4",
    "sentence-transformers>=3.4.1",
    "fastembed>=0.7.4",
    "networkx>=3.6.1",
    "pymupdf>=1.26.0",
    "pypdf>=5.4.0",
    "python-multipart>=0.0.22",
    "python-docx>=1.1.2",
    "uvicorn>=0.41.0",
]
